+++ index.html
@@ -11,61 +11,6 @@
 
 <body>
   <div id='root'></div>
-  <!-- <header>
-    <h1>뉴스스탠드</h1>
-    <span class="date">2023.02.10 금요일</span>
-  </header>
-
-  <div class="main-news">
-    <div class="news-wrapper">
-      <span>연합뉴스</span>
-      <p>[1보] ~~~</p>
-    </div>
-    <div class="news-wrapper">
-      <span>연합뉴스</span>
-      <p>[2보]~~~</p>
-    </div>
-  </div>
-
-  <section>
-    <div>
-      <nav class="view-type-wrapper">
-        <h3>전체 언론사</h3>
-        <h3>내가 구독한 언론사</h3>
-      </nav>
-      <div class="view-type-icon">
-        <span>icon</span>
-        <span>icon</span>
-      </div>
-    </div>
-
-    <ul class="news-list">
-      <li>서울경제</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>서울경제</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>서울경제</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>서울경제</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-      <li>데일리안</li>
-    </ul>
-  </section> -->
 </body>
 
 </html>
