+++ src/components/allNews/AllnewsList.js
@@ -1,0 +1,11 @@
+export default class AllNewsList {
+  constructor(name) {
+    const $component = document.createElement('li');
+    $component.innerText = name;
+
+    this.init();
+    return $component;
+  }
+
+  init() {}
+}
