+++ src/components/latestNews/index.js
